requirements.txt
nltk~=3.4.5
numpy~=1.16.0
PyYAML~=5.1.2
sklearn~=0.0
scikit-learn~=0.22.2.post1
setuptools~=41.0.1